backend/invoice_messages/requirements/base.txt
django
djangorestframework
pytest
django-cors-headers
django-webpack-loader